reorg app_gui code to make it easier to add new ui components
